Encapsulated SpecFlow transform logic
