+++ tests/unit-tests/VolleyM.Domain.Players.UnitTests/EventAssertionStepsHook.cs
@@ -9,8 +9,8 @@ namespace VolleyM.Domain.Players.UnitTests
 	/// </summary>
 	public class EventAssertionStepsHook : EventAssertionsSteps
 	{
-		public EventAssertionStepsHook(Container container, ISpecFlowTransformFactory transformFactory)
-			: base(container, transformFactory)
+		public EventAssertionStepsHook(Container container, SpecFlowTransform transform)
+			: base(container, transform)
 		{
 		}
 	}
